move local location backup setting
