+++ src/panels/config/backup/ha-config-backup-overview.ts
@@ -1,8 +1,7 @@
-import { mdiDotsVertical, mdiHarddisk, mdiPlus, mdiUpload } from "@mdi/js";
+import { mdiDotsVertical, mdiPlus, mdiUpload } from "@mdi/js";
 import type { CSSResultGroup, TemplateResult } from "lit";
 import { css, html, LitElement, nothing } from "lit";
 import { customElement, property } from "lit/decorators";
-import { isComponentLoaded } from "../../../common/config/is_component_loaded";
 import { fireEvent } from "../../../common/dom/fire_event";
 import { shouldHandleRequestSelectedEvent } from "../../../common/mwc/handle-request-selected-event";
 import "../../../components/ha-button";
@@ -33,7 +32,6 @@ import "./components/overview/ha-backup-overview-settings";
 import "./components/overview/ha-backup-overview-summary";
 import { showBackupOnboardingDialog } from "./dialogs/show-dialog-backup_onboarding";
 import { showGenerateBackupDialog } from "./dialogs/show-dialog-generate-backup";
-import { showLocalBackupLocationDialog } from "./dialogs/show-dialog-local-backup-location";
 import { showNewBackupDialog } from "./dialogs/show-dialog-new-backup";
 import { showUploadBackupDialog } from "./dialogs/show-dialog-upload-backup";
 
@@ -61,14 +59,6 @@ class HaConfigBackupOverview extends LitElement {
     }
 
     await showUploadBackupDialog(this, {});
-  }
-
-  private async _changeLocalLocation(ev) {
-    if (!shouldHandleRequestSelectedEvent(ev)) {
-      return;
-    }
-
-    showLocalBackupLocationDialog(this, {});
   }
 
   private _handleOnboardingButtonClick(ev) {
@@ -134,8 +124,6 @@ class HaConfigBackupOverview extends LitElement {
     const backupInProgress =
       "state" in this.manager && this.manager.state === "in_progress";
 
-    const isHassio = isComponentLoaded(this.hass, "hassio");
-
     return html`
       <hass-subpage
         back-path="/config/system"
@@ -143,34 +131,17 @@ class HaConfigBackupOverview extends LitElement {
         .narrow=${this.narrow}
         .header=${"Backup"}
       >
-        <div slot="toolbar-icon">
-          <ha-button-menu>
-            <ha-icon-button
-              slot="trigger"
-              .label=${this.hass.localize("ui.common.menu")}
-              .path=${mdiDotsVertical}
-            ></ha-icon-button>
-            ${isHassio
-              ? html`<ha-list-item
-                  graphic="icon"
-                  @request-selected=${this._changeLocalLocation}
-                >
-                  <ha-svg-icon
-                    slot="graphic"
-                    .path=${mdiHarddisk}
-                  ></ha-svg-icon>
-                  Change local location
-                </ha-list-item>`
-              : nothing}
-            <ha-list-item
-              graphic="icon"
-              @request-selected=${this._uploadBackup}
-            >
-              <ha-svg-icon slot="graphic" .path=${mdiUpload}></ha-svg-icon>
-              Upload backup
-            </ha-list-item>
-          </ha-button-menu>
-        </div>
+        <ha-button-menu slot="toolbar-icon">
+          <ha-icon-button
+            slot="trigger"
+            .label=${this.hass.localize("ui.common.menu")}
+            .path=${mdiDotsVertical}
+          ></ha-icon-button>
+          <ha-list-item graphic="icon" @request-selected=${this._uploadBackup}>
+            <ha-svg-icon slot="graphic" .path=${mdiUpload}></ha-svg-icon>
+            Upload backup
+          </ha-list-item>
+        </ha-button-menu>
         <div class="content">
           ${backupInProgress
             ? html`
