+++ src/panels/config/backup/ha-config-backup-settings.ts
@@ -1,6 +1,7 @@
 import type { PropertyValues } from "lit";
 import { css, html, LitElement, nothing } from "lit";
 import { customElement, property, state } from "lit/decorators";
+import { mdiDotsVertical, mdiHarddisk } from "@mdi/js";
 import { isComponentLoaded } from "../../../common/config/is_component_loaded";
 import { fireEvent } from "../../../common/dom/fire_event";
 import { debounce } from "../../../common/util/debounce";
@@ -20,6 +21,8 @@ import type { BackupConfigData } from "./components/config/ha-backup-config-data
 import "./components/config/ha-backup-config-encryption-key";
 import "./components/config/ha-backup-config-schedule";
 import type { BackupConfigSchedule } from "./components/config/ha-backup-config-schedule";
+import { shouldHandleRequestSelectedEvent } from "../../../common/mwc/handle-request-selected-event";
+import { showLocalBackupLocationDialog } from "./dialogs/show-dialog-local-backup-location";
 
 @customElement("ha-config-backup-settings")
 class HaConfigBackupSettings extends LitElement {
@@ -93,6 +96,28 @@ class HaConfigBackupSettings extends LitElement {
         .narrow=${this.narrow}
         .header=${"Automatic backups"}
       >
+        ${isComponentLoaded(this.hass, "hassio")
+          ? html`
+              <ha-button-menu slot="toolbar-icon">
+                <ha-icon-button
+                  slot="trigger"
+                  .label=${this.hass.localize("ui.common.menu")}
+                  .path=${mdiDotsVertical}
+                ></ha-icon-button>
+                <ha-list-item
+                  graphic="icon"
+                  @request-selected=${this._changeLocalLocation}
+                >
+                  <ha-svg-icon
+                    slot="graphic"
+                    .path=${mdiHarddisk}
+                  ></ha-svg-icon>
+                  Change default action location
+                </ha-list-item>
+              </ha-button-menu>
+            `
+          : nothing}
+
         <div class="content">
           <ha-card id="schedule">
             <div class="card-header">Automatic backups</div>
@@ -155,6 +180,14 @@ class HaConfigBackupSettings extends LitElement {
         </div>
       </hass-subpage>
     `;
+  }
+
+  private async _changeLocalLocation(ev) {
+    if (!shouldHandleRequestSelectedEvent(ev)) {
+      return;
+    }
+
+    showLocalBackupLocationDialog(this, {});
   }
 
   private _scheduleConfigChanged(ev) {
